+++ src/main/java/Team5/Fitnesspeaker/AlexaCommunication/Handlers/FoodHandlers/AddFoodIntentHandler.java
@@ -51,7 +51,7 @@ public class AddFoodIntentHandler implements RequestHandler {
 		//add food
 		try {
 			Pair<SearchResults, Portion> p=PortionSearchEngine.PortionSearch
-					(added_food, units, Type.FOOD, Double.valueOf(amount.intValue()).doubleValue(),db.DBGetEmail());
+					(added_food, units, Type.FOOD, amount.intValue(),db.DBGetEmail());
 			 /*if there was a search error, i.e. the food wasn't found, notify the user to about
 			 * the option of custom meal
 			 **/
